Force 32-bit integer for data stream that expects it

+++ catkit2/services/newport_picomotor/newport_picomotor.py
@@ -81,7 +81,7 @@ class NewportPicomotor(Service):
 
         # Update current position data stream.
         stream = self.axis_current_positions[axis_name]
-        stream.submit_data(np.array([current_position]))
+        stream.submit_data(np.array([current_position], dtype=np.int32))
 
         return current_position
